requirements.txt
gdal==3.9.2
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic-extra-types==2.9.0
fastapi==0.115.0
matplotlib==3.9.2
geojson-pydantic==1.1.1

uvicorn==0.31.0
gunicorn==23.0.0
prometheus-client==0.21.0
starlette_exporter==0.23.0
structlog==24.4.0
asgi-correlation-id==4.3.3
content-negotiation==1.1.2

numpy<2

# dev tooling
pytest==8.3.3
pytest-watch==4.2.0
ruff==0.6.9
mypy==1.11.2
respx==0.21.1


xarray==2024.10.0
zarr==2.18.3
fsspec==2024.10.0
odc-stac==0.3.10
pystac_client==0.8.5
rioxarray==0.18.1
boto3==1.35.66
copernicusmarine==1.3.5
